requirements.txt
Pillow==12.0.0
python-dotenv==1.2.1
python_vlc==3.0.21203
Requests==2.32.5
urllib3==1.26.18
